Update README.md - repo also depends on tactile_image_processing
Through trying to use the repo, I discovered that it also depends in tactile_image_processing and have appended it to the list on the readme.

Changed file name for launching servo control. Not sure that file I've corrected to is the desired file, but the one listed did not exist in that directory.